refactor(operation): étendre Home/Operation.vue avec une prop readOnly au lieu d'un composant Stats/Operation.vue dédié
OperationList accepte aussi readOnly et la transmet au renderer. Les drilldowns Stats (PieByCategorie, TopCategories, CategoryHeatmap) et TopOperations passent read-only=true ; la checkbox CheckOp et le bouton ✏️ sont alors masqués.
