+++ front/src/components/Stats/TopCategories.vue
@@ -8,6 +8,7 @@
       v-if="selectedCatId"
       class="top-categories__operation-list"
       :OperationRenderer="OperationRenderer"
+      :read-only="true"
     />
   </div>
 </template>
@@ -17,7 +18,7 @@
   import { useStore } from 'vuex'
   import Highcharts from 'highcharts'
   import OperationList from '../OperationList.vue'
-  import OperationRenderer from './Operation.vue'
+  import OperationRenderer from '../Home/Operation.vue'
 
   type TopCategoryItem = { IDcat: number; libelle: string; total: number }
